Implement plugin api for logging snapshots
