+++ packages/opni/opni/charts/crds/crds.yaml
@@ -43,7 +43,7 @@
 ---
 {apiVersion: apiextensions.k8s.io/v1, kind: CustomResourceDefinition, metadata: {annotations: {controller-gen.kubebuilder.io/version: v0.12.1}, name: rolebindings.monitoring.opni.io}, spec: {group: monitoring.opni.io, names: {kind: RoleBinding, listKind: RoleBindingList, plural: rolebindings, singular: rolebinding}, scope: Namespaced, versions: [{name: v1beta1, schema: {openAPIV3Schema: {properties: {apiVersion: {type: string}, kind: {type: string}, metadata: {type: object}, spec: {properties: {id: {type: string}, metadata: {properties: {resourceVersion: {type: string}}, type: object}, roleId: {type: string}, subjects: {items: {type: string}, type: array}, taints: {items: {type: string}, type: array}}, type: object}}, type: object}}, served: true, storage: true}]}}
 ---
-{apiVersion: apiextensions.k8s.io/v1, kind: CustomResourceDefinition, metadata: {annotations: {controller-gen.kubebuilder.io/version: v0.13.0}, name: recurringsnapshots.logging.opni.io}, spec: {group: logging.opni.io, names: {kind: RecurringSnapshot, listKind: RecurringSnapshotList, plural: recurringsnapshots, singular: recurringsnapshot}, scope: Namespaced, versions: [{additionalPrinterColumns: [{jsonPath: .metadata.creationTimestamp, name: Age, type: date}, {jsonPath: .status.state, name: State, type: boolean}], name: v1beta1, schema: {openAPIV3Schema: {properties: {apiVersion: {type: string}, kind: {type: string}, metadata: {type: object}, spec: {properties: {creation: {properties: {cronSchedule: {type: string}, timeLimit: {type: string}}, required: [cronSchedule], type: object}, retention: {properties: {maxAge: {type: string}, maxCount: {type: integer}}, type: object}, snapshot: {properties: {allowPartial: {type: boolean}, ignoreUnavailable: {type: boolean}, includeGlobalState: {type: boolean}, indices: {items: {type: string}, type: array}, repository: {properties: {name: {type: string}}, type: object, x-kubernetes-map-type: atomic}}, required: [repository], type: object}}, required: [creation, snapshot], type: object}, status: {properties: {executionStatus: {properties: {cause: {type: string}, lastExecution: {format: date-time, type: string}, message: {type: string}, status: {type: string}}, type: object}, state: {type: string}}, type: object}}, type: object}}, served: true, storage: true, subresources: {status: {}}}]}}
+{apiVersion: apiextensions.k8s.io/v1, kind: CustomResourceDefinition, metadata: {annotations: {controller-gen.kubebuilder.io/version: v0.13.0}, name: recurringsnapshots.logging.opni.io}, spec: {group: logging.opni.io, names: {kind: RecurringSnapshot, listKind: RecurringSnapshotList, plural: recurringsnapshots, singular: recurringsnapshot}, scope: Namespaced, versions: [{additionalPrinterColumns: [{jsonPath: .metadata.creationTimestamp, name: Age, type: date}, {jsonPath: .status.state, name: State, type: boolean}], name: v1beta1, schema: {openAPIV3Schema: {properties: {apiVersion: {type: string}, kind: {type: string}, metadata: {type: object}, spec: {properties: {creation: {properties: {cronSchedule: {type: string}, timeLimit: {type: string}}, required: [cronSchedule], type: object}, retention: {properties: {maxAge: {type: string}, maxCount: {format: int32, type: integer}}, type: object}, snapshot: {properties: {allowPartial: {type: boolean}, ignoreUnavailable: {type: boolean}, includeGlobalState: {type: boolean}, indices: {items: {type: string}, type: array}, repository: {properties: {name: {type: string}}, type: object, x-kubernetes-map-type: atomic}}, required: [repository], type: object}}, required: [creation, snapshot], type: object}, status: {properties: {executionStatus: {properties: {cause: {type: string}, lastExecution: {format: date-time, type: string}, message: {type: string}, status: {type: string}}, type: object}, state: {type: string}}, type: object}}, type: object}}, served: true, storage: true, subresources: {status: {}}}]}}
 ---
 {apiVersion: apiextensions.k8s.io/v1, kind: CustomResourceDefinition, metadata: {annotations: {controller-gen.kubebuilder.io/version: v0.13.0}, name: rolebindings.monitoring.opni.io}, spec: {group: monitoring.opni.io, names: {kind: RoleBinding, listKind: RoleBindingList, plural: rolebindings, singular: rolebinding}, scope: Namespaced, versions: [{name: v1beta1, schema: {openAPIV3Schema: {properties: {apiVersion: {type: string}, kind: {type: string}, metadata: {type: object}, spec: {properties: {id: {type: string}, metadata: {properties: {resourceVersion: {type: string}}, type: object}, roleId: {type: string}, subjects: {items: {type: string}, type: array}, taints: {items: {type: string}, type: array}}, type: object}}, type: object}}, served: true, storage: true}]}}
 ---
