+++ plugins/logging/pkg/gateway/drivers/management/cluster_driver.go
@@ -17,6 +17,11 @@ type ClusterDriver interface {
 	UpgradeAvailable(ctx context.Context, opniVersion string) (bool, error)
 	DoUpgrade(ctx context.Context, opniVersion string) error
 	GetStorageClasses(context.Context) ([]string, error)
+	CreateOrUpdateSnapshot(ctx context.Context, snapshot *loggingadmin.Snapshot) error
+	// Get snapshot only returns recurring snapshots because oneoff snapshots are immutable once created
+	GetRecurringSnapshot(ctx context.Context, ref *loggingadmin.SnapshotReference) (*loggingadmin.Snapshot, error)
+	DeleteSnapshot(ctx context.Context, ref *loggingadmin.SnapshotReference) error
+	ListAllSnapshots(ctx context.Context) (*loggingadmin.SnapshotStatusList, error)
 }
 
 var Drivers = driverutil.NewDriverCache[ClusterDriver]()
